+++ branches/psycle-186/psycle/src/psycle/host/WaveOut.cpp
@@ -202,7 +202,7 @@ namespace psycle
 				Stop();
 			}
 			_capPorts.push_back(port);
-			_portMapping[idx]=_capPorts.size()-1;
+			_portMapping[idx]=(int)(_capPorts.size()-1);
 			if (isplaying)
 			{
 				return Start();
